optionally do not delete

+++ index.js
@@ -7,12 +7,13 @@ function tmpdir () {
   return os.tmpdir ? os.tmpdir() : process.env.TMPDIR
 }
 
-module.exports = function () {
-
+module.exports = function (opts) {
+  opts = opts || {}
   return function (name, opts, cb) {
     name = name || 'db_' + Date.now()
     var dir = path.join(tmpdir(), name)
-    rimraf.sync(dir)
+    if(opts.clean !== false)
+      rimraf.sync(dir)
     return level(dir, opts, cb)
   }
